+++ src/main/java/so/alaz/cleansingpotions/platform/services/IPlatformHelper.java
@@ -8,6 +8,8 @@ public interface IPlatformHelper {
 
     String getPlatformName();
 
+    String getModVersion();
+
     boolean isModLoaded(String modId);
 
     boolean isDevelopmentEnvironment();